Major Refactor of Combat

+++ ThirtyDayHero/Item/Armors/IArmor.cs
@@ -4,6 +4,6 @@
     {
         ArmorType ArmorType { get; }
 
-        uint GetReducedDamage(uint damageAmount, DamageType damageType);
+        float GetReducedDamage(float damageAmount, DamageType damageType);
     }
 }
